Move environments into a Environment/ subdirectory.
